requirements.txt
solana
solders
aiohttp
python-dotenv
anchorpy
construct
based58
websockets
telethon
aiofiles
pandas
aiosqlite
python-telegram-bot